Split Skills into separate component
